add file_test.go
